👌 Address Samuel's comments
